SummonersRiftDefault is now EmptyMapScript, which allows any map to be initialized in a barebones state
