+++ GameServerLib/Scripting/CSharp/EmptyMapScript.cs
@@ -1,0 +1,149 @@
+using System;
+using System.Collections.Generic;
+using System.Numerics;
+using GameServerCore.Domain;
+using GameServerCore.Domain.GameObjects;
+using GameServerCore.Enums;
+using GameServerCore.Maps;
+using LeagueSandbox.GameServer;
+using LeagueSandbox.GameServer.GameObjects;
+using LeagueSandbox.GameServer.GameObjects.AttackableUnits.AI;
+using LeagueSandbox.GameServer.GameObjects.Other;
+using LeagueSandbox.GameServer.Maps;
+
+namespace MapScripts
+{
+    public class EmptyMapScript : IMapScript
+    {
+        public bool HasInnerTurrets { get; set; } = true;
+
+        //General Map variable
+        private IMap _map;
+
+        //Stuff about minions
+        public bool SpawnEnabled { get; set; }
+        public long FirstSpawnTime { get; set; } = 30 * 1000;
+        public long NextSpawnTime { get; set; } = 30 * 1000;
+        public long SpawnInterval { get; set; } = 30 * 1000;
+        public bool MinionPathingOverride { get; set; } = true;
+
+        //General things that will affect players globaly, such as default gold per-second, Starting gold....
+        public float GoldPerSecond { get; set; } = 1.9f;
+        public float StartingGold { get; set; } = 475.0f;
+        public bool HasFirstBloodHappened { get; set; } = false;
+        public bool IsKillGoldRewardReductionActive { get; set; } = true;
+        public int BluePillId { get; set; } = 2001;
+        public long FirstGoldTime { get; set; } = 90 * 1000;
+
+        //Tower type enumeration might vary slightly from map to map, so we set that up here
+        public TurretType GetTurretType(int trueIndex, LaneID lane, TeamId teamId)
+        {
+            return TurretType.OUTER_TURRET;
+        }
+
+        //List of each turret model present in this map, being organized between team and tower type
+        public Dictionary<TeamId, Dictionary<TurretType, string>> TowerModels { get; set; } = new Dictionary<TeamId, Dictionary<TurretType, string>>
+        {
+            {TeamId.TEAM_BLUE, new Dictionary<TurretType, string>
+            {
+                {TurretType.OUTER_TURRET, "OrderTurretNormal" },
+            } },
+            {TeamId.TEAM_PURPLE, new Dictionary<TurretType, string>
+            {
+                {TurretType.OUTER_TURRET, "ChaosTurretWorm" },
+            } }
+        };
+
+        //Turret Items
+        public Dictionary<TurretType, int[]> TurretItems { get; set; } = new Dictionary<TurretType, int[]>
+        {
+            { TurretType.OUTER_TURRET, new[] { 1500, 1501, 1502, 1503 } },
+        };
+
+        //List of every path minions will take, separated by team and lane
+        public Dictionary<LaneID, List<Vector2>> MinionPaths { get; set; } = new Dictionary<LaneID, List<Vector2>>();
+
+        //List of every wave type
+        public Dictionary<string, List<MinionSpawnType>> MinionWaveTypes = new Dictionary<string, List<MinionSpawnType>>
+        { {"RegularMinionWave", new List<MinionSpawnType>
+        {
+            MinionSpawnType.MINION_TYPE_MELEE,
+            MinionSpawnType.MINION_TYPE_MELEE,
+            MinionSpawnType.MINION_TYPE_MELEE,
+            MinionSpawnType.MINION_TYPE_CASTER,
+            MinionSpawnType.MINION_TYPE_CASTER,
+            MinionSpawnType.MINION_TYPE_CASTER }
+        }
+        };
+
+        //Here you setup the conditions of which wave will be spawned
+        public Tuple<int, List<MinionSpawnType>> MinionWaveToSpawn(float gameTime, int cannonMinionCount, bool isInhibitorDead, bool areAllInhibitorsDead)
+        {
+            return new Tuple<int, List<MinionSpawnType>>(0, MinionWaveTypes["RegularMinionWave"]);
+        }
+
+        //Minion models for this map
+        public Dictionary<TeamId, Dictionary<MinionSpawnType, string>> MinionModels { get; set; } = new Dictionary<TeamId, Dictionary<MinionSpawnType, string>>
+        {
+            {TeamId.TEAM_BLUE, new Dictionary<MinionSpawnType, string>{
+                {MinionSpawnType.MINION_TYPE_MELEE, "Blue_Minion_Basic"},
+                {MinionSpawnType.MINION_TYPE_CASTER, "Blue_Minion_Wizard"},
+            }},
+            {TeamId.TEAM_PURPLE, new Dictionary<MinionSpawnType, string>{
+                {MinionSpawnType.MINION_TYPE_MELEE, "Red_Minion_Basic"},
+                {MinionSpawnType.MINION_TYPE_CASTER, "Red_Minion_Wizard"},
+            }}
+        };
+
+        //This function is executed in-between Loading the map structures and applying the structure protections. Is the first thing on this script to be executed
+        public void Init(IMap map)
+        {
+            _map = map;
+
+            SpawnEnabled = map.IsMinionSpawnEnabled();
+            map.AddSurrender(1200000.0f, 300000.0f, 30.0f);
+
+        }
+
+        //This function gets executed every server tick
+        public void Update(float diff)
+        {
+        }
+
+
+        public float GetGoldFor(IAttackableUnit u)
+        {
+            return 0;
+        }
+
+        public float GetExperienceFor(IAttackableUnit u)
+        {
+            return 0.0f;
+        }
+
+        public void SetMinionStats(ILaneMinion m)
+        {
+            // Same for all minions
+            m.Stats.MoveSpeed.BaseValue = 325.0f;
+
+            switch (m.MinionSpawnType)
+            {
+                case MinionSpawnType.MINION_TYPE_MELEE:
+                    m.Stats.CurrentHealth = 475.0f;
+                    m.Stats.HealthPoints.BaseValue = 475.0f;
+                    m.Stats.AttackDamage.BaseValue = 12.0f;
+                    m.Stats.Range.BaseValue = 180.0f;
+                    m.Stats.AttackSpeedFlat = 1.250f;
+                    m.IsMelee = true;
+                    break;
+                case MinionSpawnType.MINION_TYPE_CASTER:
+                    m.Stats.CurrentHealth = 279.0f;
+                    m.Stats.HealthPoints.BaseValue = 279.0f; ;
+                    m.Stats.AttackDamage.BaseValue = 23.0f;
+                    m.Stats.Range.BaseValue = 600.0f;
+                    m.Stats.AttackSpeedFlat = 0.670f;
+                    break;
+            }
+        }
+    }
+}
